Moved GReaderColumns into a separate file, renamed into dt::read::Columns
