+++ Makefile
@@ -444,6 +444,7 @@ fast_objects = $(addprefix $(BUILDDIR)/, \
 	python/long.o             \
 	python/string.o           \
 	read/column.o             \
+	read/columns.o          \
 	rowindex.o                \
 	rowindex_array.o          \
 	rowindex_slice.o          \
@@ -630,7 +631,7 @@ $(BUILDDIR)/csv/py_csv.h: c/csv/py_csv.h $(BUILDDIR)/py_utils.h
 	@echo • Refreshing c/csv/py_csv.h
 	@cp c/csv/py_csv.h $@
 
-$(BUILDDIR)/csv/reader.h: c/csv/reader.h $(BUILDDIR)/column.h $(BUILDDIR)/read/column.h $(BUILDDIR)/datatable.h $(BUILDDIR)/memrange.h $(BUILDDIR)/utils/array.h $(BUILDDIR)/utils/pyobj.h $(BUILDDIR)/utils/shared_mutex.h $(BUILDDIR)/writebuf.h
+$(BUILDDIR)/csv/reader.h: c/csv/reader.h $(BUILDDIR)/column.h $(BUILDDIR)/read/column.h $(BUILDDIR)/read/columns.h $(BUILDDIR)/datatable.h $(BUILDDIR)/memrange.h $(BUILDDIR)/utils/array.h $(BUILDDIR)/utils/pyobj.h $(BUILDDIR)/utils/shared_mutex.h $(BUILDDIR)/writebuf.h
 	@echo • Refreshing c/csv/reader.h
 	@cp c/csv/reader.h $@
 
@@ -676,6 +677,10 @@ $(BUILDDIR)/python/string.h: c/python/string.h $(BUILDDIR)/utils/pyobj.h
 $(BUILDDIR)/read/column.h: c/read/column.h $(BUILDDIR)/memrange.h $(BUILDDIR)/utils/pyobj.h $(BUILDDIR)/writebuf.h
 	@echo • Refreshing c/read/column.h
 	@cp c/read/column.h $@
+
+$(BUILDDIR)/read/columns.h: c/read/columns.h $(BUILDDIR)/read/column.h
+	@echo • Refreshing c/read/columns.h
+	@cp c/read/columns.h $@
 
 
 $(BUILDDIR)/utils/alloc.h: c/utils/alloc.h
@@ -916,9 +921,15 @@ $(BUILDDIR)/python/string.o : c/python/string.cc $(BUILDDIR)/python/string.h $(B
 	@echo • Compiling $<
 	@$(CC) -c $< $(CCFLAGS) -o $@
 
+
 $(BUILDDIR)/read/column.o : c/read/column.cc $(BUILDDIR)/read/column.h $(BUILDDIR)/csv/reader.h $(BUILDDIR)/csv/reader_parsers.h $(BUILDDIR)/python/string.h
 	@echo • Compiling $<
 	@$(CC) -c $< $(CCFLAGS) -o $@
+
+$(BUILDDIR)/read/columns.o : c/read/columns.cc $(BUILDDIR)/read/columns.h $(BUILDDIR)/csv/reader_parsers.h
+	@echo • Compiling $<
+	@$(CC) -c $< $(CCFLAGS) -o $@
+
 
 $(BUILDDIR)/rowindex.o : c/rowindex.cc $(BUILDDIR)/rowindex.h $(BUILDDIR)/utils.h $(BUILDDIR)/utils/assert.h $(BUILDDIR)/utils/omp.h
 	@echo • Compiling $<
